// src/lib/api/generated/models/exportLimits.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.11
 */

export interface ExportLimits {
  /**
     * @minimum 0
     * @nullable
     */
  max_items?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  max_output_bytes?: number | null;
}