// src/lib/api/generated/models/taskStatus.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.1.0
 */

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  queued: 'queued',
  validating: 'validating',
  running: 'running',
  succeeded: 'succeeded',
  failed: 'failed',
  partially_succeeded: 'partially_succeeded',
  cancelled: 'cancelled',
} as const;